requirements.txt
flask>=3.1.0
openai>=2.6.0
python-dotenv>=1.1.0
